rename ns
